+++ OpenAirMarket-ejb/src/test/java/com/structureeng/persistence/history/TenantHistoryListenerTest.java
@@ -155,7 +155,7 @@ public class TenantHistoryListenerTest extends AbstractPersistenceTest {
         return Company.newBuilder().setReferenceId(referenceId).setName(name).build();
     }
 
-    private Tenant createTenant(long id) {
+    private Tenant createTenant(int id) {
         Tenant tenant = new Tenant();
         tenant.setId(id);
         return tenant;
